feat: Automate creating amalgamation distribution
We enhance the build script (batch and bash), so it can generate
amalgamation distributions for both the UNIX and Windows platforms.

To generate a distribution for a specific version, first use make to
list all tags and check out the specified tag in the project root
directory:

  $ make taglist
  $ make co TAG=<tag>

Then go to the tools directory and run the script to generate tar.gz
for UNIX and generate zip for Windows:

  $ cd .tag/<tag>/tools
  $ ./build cleanall
  $ ./build tar
  $ ./build zip

Notice that this enhancement was added after the official release
v3.1.5.1c1bb23, so you need to specify a tag newer than v3.1.5.

resolves #43